Make some headway on removing deps on `lib/auth`
