+++ lib/auth/authclient/http_client.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package auth
+package authclient
 
 import (
 	"context"
